+++ lib/alf.rb
@@ -77,7 +77,7 @@ module Alf
   # 
   class Environment
     require 'alf/environment/class_methods'
-    require 'alf/environment/base'
+    require 'alf/environment/instance_methods'
     require 'alf/environment/explicit'
     require 'alf/environment/folder'